Carrinho funcionando

+++ app/src/main/java/br/com/lotus_projeto_integrador/lotus/CarrinhoProduto.java
@@ -35,7 +35,7 @@ public class CarrinhoProduto {
         this.valor = valor;
     }
 
-    public CarrinhoProduto(int idProduto,String Nome,int quantidade,int valor) {
+    public CarrinhoProduto(int idProduto,String Nome,int quantidade,double valor) {
         this.Nome = Nome;
         this.IdProduto = idProduto;
         this.quantidade = quantidade;
